Move get_next_url into apps.common
