+++ apps/common/__init__.py
@@ -6,6 +6,8 @@ from decimal import Decimal
 from os import path
 from pathlib import Path
 from textwrap import wrap
+from typing import overload
+from urllib.parse import urljoin, urlparse, urlunparse
 
 import pendulum
 from decorator import decorator
@@ -13,7 +15,9 @@ from flask import (
     abort,
     render_template,
     render_template_string,
+    request,
     session,
+    url_for,
 )
 from flask import (
     current_app as app,
@@ -376,3 +380,30 @@ def render_markdown(source, template="about/template.html", **view_variables):
 
     view_variables.update(content=content, title=metadata["title"])
     return render_template(page_template(metadata, template), **view_variables)
+
+
+def make_safe_url(target: str) -> str | None:
+    ref_url = urlparse(request.host_url)
+    test_url = urlparse(urljoin(request.host_url, target))
+    if test_url.scheme in ("http", "https") and ref_url.netloc == test_url.netloc:
+        return urlunparse(test_url)
+    return None
+
+
+@overload
+def get_next_url(default: str) -> str: ...
+
+
+@overload
+def get_next_url(default: None) -> str | None: ...
+
+
+def get_next_url(default=None):
+    next_url = request.args.get("next")
+    if next_url:
+        if safe_url := make_safe_url(next_url):
+            return safe_url
+        app.logger.error(f"Dropping unsafe next URL {repr(next_url)}")
+    if default is None:
+        default = url_for(".account")
+    return default
